wayland: refactor to remove WaylandContext
