+++ src/platform/linux/mod.rs
@@ -332,8 +332,8 @@ impl EventsLoop {
     }
 
     pub fn new_wayland() -> Result<EventsLoop, ()> {
-        wayland::WaylandContext::init()
-            .map(|ctx| EventsLoop::Wayland(wayland::EventsLoop::new(ctx)))
+        wayland::EventsLoop::new()
+            .map(EventsLoop::Wayland)
             .ok_or(())
     }
 
@@ -347,7 +347,7 @@ impl EventsLoop {
     #[inline]
     pub fn get_available_monitors(&self) -> VecDeque<MonitorId> {
         match *self {
-            EventsLoop::Wayland(ref evlp) => wayland::get_available_monitors(evlp.context())
+            EventsLoop::Wayland(ref evlp) => evlp.get_available_monitors()
                                     .into_iter()
                                     .map(MonitorId::Wayland)
                                     .collect(),
@@ -361,7 +361,7 @@ impl EventsLoop {
     #[inline]
     pub fn get_primary_monitor(&self) -> MonitorId {
         match *self {
-            EventsLoop::Wayland(ref evlp) => MonitorId::Wayland(wayland::get_primary_monitor(evlp.context())),
+            EventsLoop::Wayland(ref evlp) => MonitorId::Wayland(evlp.get_primary_monitor()),
             EventsLoop::X(ref evlp) => MonitorId::X(x11::get_primary_monitor(evlp.x_connection())),
         }
     }
